requirements.txt
streamlit 
pandas 
google.generativeai
python-dotenv
googletrans==4.0.0-rc1
pillow
pillow-heif
streamlit-authenticator